pequenos anadidos

+++ lib/tareas_loops/tareas.cpp
@@ -80,14 +80,14 @@ struct __attribute__((packed)) Envio {
   float teta_ref; //4
   float v_der; //5
   float v_izq; //6
-  int32_t v_der_ref;
-  int32_t v_izq_ref;
-  int32_t v_total;
-  int32_t v_total_ref;
-  int32_t x_pos;
-  int32_t y_pos;
-  int32_t x_ref;
-  int32_t y_ref;  // hasta aca parecen haber 57 bytes
+  float v_der_ref;
+  float v_izq_ref;
+  float v_total;
+  float v_total_ref;
+  float x_pos;
+  float y_pos;
+  float x_ref;
+  float y_ref;  // hasta aca parecen haber 57 bytes
 
   //uint8_t checksum;      // Para validar integridad
 };
